Perdaryta studentų vardų ir pavardžių įvesis

+++ array/calc.cpp
@@ -1,19 +1,19 @@
 #include "calc.h"
 #include <algorithm>
 
-float galutinisVid(const Studentas &A){
+double galutinisVid(const Studentas &A){
     if(A.ndKiekis==0) return 0.6*A.rez;
 
     int sum = 0;
     for(int x = 0; x < A.ndKiekis; x++) // is A.nd masyvo prideda kiekviena nari.
         sum += A.nd[x];
 
-    float vid = (float)sum / A.ndKiekis;
+    double vid = (double)sum / A.ndKiekis;
     return 0.4 * vid + 0.6 * A.rez;
 }
 
 
-float galutinisMed(const Studentas &A){
+double galutinisMed(const Studentas &A){
     if(A.ndKiekis==0) return 0.6*A.rez;
 
     int* nd = new int[A.ndKiekis];
@@ -23,13 +23,12 @@ float galutinisMed(const Studentas &A){
 
     std::sort(nd, nd + A.ndKiekis);
 
-    float med =
+    double med =
         (A.ndKiekis % 2 == 0) // jei lyginis, tai dvieju viduriniu nd vektoriaus nariu mediana paskaiciuoja
         ? (nd[A.ndKiekis/2] + nd[A.ndKiekis/2 - 1]) / 2.0
         : nd[A.ndKiekis/2];
 
-    float rezultatas = 0.4 * med + 0.6 * A.rez; // naudojama f, kad skaiciavimai butu atliekami su grynais float'ais, o ne double, nes paprastai 0.4 laikomas double. Beje, mum nereikia tokiu tiksliu vidurkiu, kaip double leistu, taigi sutaupom atminties. 
-    
+    double rezultatas = 0.4 * med + 0.6 * A.rez;
     delete[] nd;
     
     return rezultatas;
